Address CVE-2018-13988
Poppler through 0.62 contains an out of bounds read vulnerability due to an
incorrect memory access that is not mapped in its memory space, as
demonstrated by pdfunite. This can result in memory corruption and denial
of service. This may be exploitable when a victim opens a specially crafted
PDF file.

Signed-off-by: Peter O'Connor <peter@solus-project.com>